CRUD PO

+++ src/main/java/com/umirrosi/springrestapi/repository/PurchaseOrderDetailRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PurchaseOrderDetailRepo extends JpaRepository<PurchaseOrderDetailEntity, Integer> {
+public interface PurchaseOrderDetailRepo extends JpaRepository<PurchaseOrderDetailEntity, Long> {
 }
