+++ src/main/java/com/umirrosi/springrestapi/repository/PurchaseOrderRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PurchaseOrderRepo extends JpaRepository<PurchaseOrderEntity, Integer> {
+public interface PurchaseOrderRepo extends JpaRepository<PurchaseOrderEntity, Long> {
 }
