+++ src/main/java/com/umirrosi/springrestapi/repository/SupplierRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupplierRepo extends JpaRepository<SupplierEntity, Integer> {
+public interface SupplierRepo extends JpaRepository<SupplierEntity, Long> {
 }
